Upgrades to support DraftJS v0.11.x
To support DraftJS v0.11.x forward, handleBeforeInput needs to take an
additional parameter `eventTimestamp` for synchronizing collaborative
actions. This change adds support for eventTimestamp, updates all tests
and snapshots, and updates the peerDependencies to match.

BREAKING CHANGE: The changes to DraftJS and draft-js-plugin-editor are
not backwards compatible with the 0.10.x/2.0.0 versions respectively.
The peerDependencies were updated to reflect this and a note was added
to the readme.